Add example of how to disable autoindent to vimrc
I was dissatisfied with some of default indentation rules for ruby so
I found how to disable them. Unfortunately, it disables all of them,
so I've decided to not use it. Still, it's a useful thing that I would
like to keep here commented out, mostly as reminder for future.